refactor(ui,workbench): move thread grouping and repository label to platform-neutral ui modules (CODE-276)
