+++ packages/ui/src/native/index.ts
@@ -1,4 +1,6 @@
 export { AGENT_INITIALS, AGENT_LABELS } from '../agent-meta';
+export { repositoryLabel } from '../repository-label';
+export * from '../thread-groups';
 export * from './agent-icon';
 export * from './empty-state';
 export * from './mobile-home';
